Allow commas in config
User federation config is a MultivaluedMap. A multivalued map
has a key and a list of values.

The correct terraform config would be:
config = {
key = ["value1", "value2"]
}

Changing the config structure to array of lists would break
existing configurations therefore only the first value is
set and taken into account.